Solved AoC2021 Day 9 Part 2

+++ aoc2021/aoc2021-kotlin/src/test/kotlin/de/havox_design/aoc2021/day09/Day09Test.kt
@@ -15,7 +15,7 @@ class Day09Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Long) =
+    fun testProcessPart2(filename: String, expectedResult: Int) =
         SmokeBasin(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
